Changed the sql

+++ OneTimeProgress/BussinessEntity/Entity.cs
@@ -68,7 +68,7 @@ namespace OneTimeProgress.BussinessEntity
         public string EndTime { get; set; } //starttime + duration
         public string ActualStartTime { get; set; }
         public string ActualEndTime { get; set; }
-        public int TimeDifference { get; set; }
+        public double TimeDifference { get; set; }
         public double CurrentTimeMinusActualStartTime { get; set; } // caluclated if Status is In Progress
         public string Status { get; set; }
         public string Colour { get; set; }// status completed then check the timedifference - duration if it is>1 make it red bar
